requirements.txt
fastapi
uvicorn
python-dotenv
requests
openai
vaderSentiment
SpeechRecognition
pyttsx3
jinja2
python-multipart
google-generativeai
matplotlib
schedule
